Ядро 1.4.4
В хранилище лежала версия 1.1.5, а работа с тех пор шла только на рабочей
машине: собрать узел не мог никто, кроме неё. Дерево приводится к тому,
что действительно работает на стенде.

Из последнего, найденного при сверке остатка денежного ящика: автоизъятие
при закрытии смены собирало документ как Money(currentCash, 0), где счётчик
хранится в тиынах, а первое поле Money — тенге. При включённом автоизъятии
9 720 тенге ящика ушли бы в ОФД документом на 972 000. Исправлено на
Money.fromTiyn; проверка testCloseShiftAutoCashoutTakesWhatIsInTheDrawer
без правки падает.

+++ core-domain/src/commonMain/kotlin/io/github/texport/superkassa/core/domain/api/model/kkm/CashOperationRequest.kt
@@ -1,5 +1,7 @@
 package io.github.texport.superkassa.core.domain.api.model.kkm
 
+import io.github.texport.superkassa.core.domain.api.model.common.Decimal
+
 /**
  * Запрос на проведение операции с наличными (внесение/изъятие).
  *
@@ -9,7 +11,7 @@ package io.github.texport.superkassa.core.domain.api.model.kkm
  */
 data class CashOperationRequest(
     val pin: String = "",
-    val amount: Double,
+    val amount: Decimal,
     val idempotencyKey: String
 ) {
     /**
