+++ core-presentation/src/commonMain/kotlin/io/github/texport/superkassa/core/presentation/api/model/kkm/CashOperationRequest.kt
@@ -1,5 +1,6 @@
 package io.github.texport.superkassa.core.presentation.api.model.kkm
 
+import io.github.texport.superkassa.core.domain.api.model.common.Decimal
 import io.github.texport.superkassa.core.presentation.api.annotations.Schema
 import io.github.texport.superkassa.core.presentation.api.annotations.DecimalMin
 import io.github.texport.superkassa.core.presentation.api.annotations.NotBlank
@@ -13,7 +14,7 @@ import kotlinx.serialization.Serializable
 data class CashOperationRequest(
     @Schema(description = "Сумма операции в тенге", example = "5000.00")
     @field:DecimalMin("0.01", message = "Сумма операции должна быть положительной")
-    val amount: Double,
+    val amount: Decimal,
     @Schema(description = "Ключ идемпотентности", example = "unique-id-123")
     @field:NotBlank(message = "Ключ идемпотентности обязателен")
     val idempotencyKey: String
